fix a bug on loss computing: add weighted loss computing

+++ src/main/java/ml/shifu/shifu/core/dtrain/dt/DTMasterParams.java
@@ -47,14 +47,14 @@ public class DTMasterParams extends HaltBytable {
     private Map<Integer, TreeNode> todoNodes;
 
     /**
-     * Sum of training counts accumulated by workers.
+     * Sum of weighted training counts accumulated by workers.
      */
-    private long trainCount;
+    private double trainCount;
 
     /**
-     * Sum of validation counts accumulated by workers.
+     * Sum of weighted validation counts accumulated by workers.
      */
-    private long validationCount;
+    private double validationCount;
 
     /**
      * Sum of train error accumulated by workers.
@@ -84,7 +84,7 @@ public class DTMasterParams extends HaltBytable {
     public DTMasterParams() {
     }
 
-    public DTMasterParams(long trainCount, double trainError, long validationCount, double validationError) {
+    public DTMasterParams(double trainCount, double trainError, double validationCount, double validationError) {
         this.trainCount = trainCount;
         this.trainError = trainError;
         this.validationCount = validationCount;
@@ -128,8 +128,8 @@ public class DTMasterParams extends HaltBytable {
 
     @Override
     public void doWrite(DataOutput out) throws IOException {
-        out.writeLong(trainCount);
-        out.writeLong(validationCount);
+        out.writeDouble(trainCount);
+        out.writeDouble(validationCount);
         out.writeDouble(trainError);
         out.writeDouble(validationError);
         out.writeBoolean(this.isSwitchToNextTree);
@@ -155,8 +155,8 @@ public class DTMasterParams extends HaltBytable {
 
     @Override
     public void doReadFields(DataInput in) throws IOException {
-        this.trainCount = in.readLong();
-        this.validationCount = in.readLong();
+        this.trainCount = in.readDouble();
+        this.validationCount = in.readDouble();
         this.trainError = in.readDouble();
         this.validationError = in.readDouble();
         this.isSwitchToNextTree = in.readBoolean();
@@ -185,14 +185,14 @@ public class DTMasterParams extends HaltBytable {
     /**
      * @return the trainCount
      */
-    public long getTrainCount() {
+    public double getTrainCount() {
         return trainCount;
     }
 
     /**
      * @return the validationCount
      */
-    public long getValidationCount() {
+    public double getValidationCount() {
         return validationCount;
     }
 
@@ -200,7 +200,7 @@ public class DTMasterParams extends HaltBytable {
      * @param trainCount
      *            the trainCount to set
      */
-    public void setTrainCount(long trainCount) {
+    public void setTrainCount(double trainCount) {
         this.trainCount = trainCount;
     }
 
@@ -208,7 +208,7 @@ public class DTMasterParams extends HaltBytable {
      * @param validationCount
      *            the validationCount to set
      */
-    public void setValidationCount(long validationCount) {
+    public void setValidationCount(double validationCount) {
         this.validationCount = validationCount;
     }
